requirements.txt
django
requests
beautifulsoup4
geopy
googlemaps
django-allauth
django-htmx
python-decouple
django-filter
retrying